+++ src/main/java/LottoConfig.java
@@ -1,7 +1,7 @@
 import controller.LottoController;
-import controller.dto.LottoDtoMapper;
-import model.LottoNumberGenerator;
-import model.LottoRankCalculator;
+import controller.LottoDtoMapper;
+import model.numbers.LottoNumberGenerator;
+import model.rank.LottoRankCalculator;
 import model.LottoStore;
 import view.InputParser;
 import view.InputView;
